torrent-gui: added tracker gui
