+++ src/main/java/ru/spbau/javacourse/torrent/TrackerApp.java
@@ -1,24 +1,20 @@
 package ru.spbau.javacourse.torrent;
 
-
 import javafx.application.Application;
 import javafx.stage.Stage;
 import lombok.extern.java.Log;
-import ru.spbau.javacourse.torrent.gui.ClientGUI;
+import ru.spbau.javacourse.torrent.gui.TrackerGUI;
 
 import java.util.logging.Level;
 
-/**
- * Gui Application for Torrent Client
- */
 @Log
-public class ClientApp extends Application {
+public class TrackerApp extends Application {
 
     @Override
     public void start(Stage primaryStage) throws Exception {
-        log.log(Level.INFO, "Starting client application...");
+        log.log(Level.INFO, "Starting tracker application ...");
 
-        final ClientGUI gui = new ClientGUI(primaryStage);
+        final TrackerGUI gui = new TrackerGUI(primaryStage);
         gui.show();
     }
